[Tizen] Fix Svace issue for 64bit
Solve some convert from size_t to uint32_t case.

+++ dali/public-api/rendering/texture-set.h
@@ -2,7 +2,7 @@
 #define DALI_TEXTURE_SET_H
 
 /*
- * Copyright (c) 2020 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2022 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -154,7 +154,7 @@ public:
    * @SINCE_1_1.43
    * @return The number of textures in the TextureSet
    */
-  size_t GetTextureCount() const;
+  uint32_t GetTextureCount() const;
 
 public:
   /**
